:art: #142 duration_in_seconds は秒単位でしか取得しないので int 型に変更

+++ src/domain/repository/video_repository_interface.py
@@ -7,7 +7,7 @@ class AnalysisVideoDto(TypedDict):
 
 class AnalysisVideoResult(TypedDict):
     summary: str
-    duration_in_seconds: float
+    duration_in_seconds: int
 
 
 class VideoRepositoryInterface(Protocol):
